moving commands from `%sqlcmd` to separate files (#654)
* test

* refactor magic_cmd.py

* refactor test_magic

* fix lint refactor magic_cmd

* refactor magic cmd lint1

* add init file

* refactoring changes

* removed CmdParser class

* removed CmdParser class-1

* removed CmdParser class-2

* re-adding cmd_parser
